Drop in basic code
Contains Neopixel code to control a circular strip of LEDs, basic timeline logic for easing animation, color shift Gamma correction for LED color accuracy, and a basic state machine to program animations without use of delay (so sensors can be used and will be responsive).  This project is currently wired to a single push button to restart the animation, and will soon be hooked up to a hall sensor to allow for magnetic restart of the animation.